(senator-message): No more logging of messages on XEmacs.
(senator-find-token-before, senator-find-previous-token,
senator-find-last-token, senator-find-next-token): Removed these
no more used functions.

(senator-step-at-parent, senator-previous-token-or-parent): New
functions.

(senator-next-token, senator-previous-token): New implementation
based on `semantic-find-nonterminal-by-overlay-next' and
`semantic-find-nonterminal-by-overlay-prev'.

(senator-beginning-of-defun): Fixed the bad
`senator-step-at-start-end-token-ids' value.  Ensure that the
point is always at the beginning of the previous token found.

(senator-beginning-of-defun, senator-end-of-defun): Doc string
updated to inform that the ARG parameter is not yet handled.