(semantic-bovine-bnf-table, semantic-bnf-keyword-table): Added new TOKENTABLE and PARSERMODE keywords.
(semantic-bnf-token-name-symbol, semantic-bnf-token-token-type,
semantic-bnf-token-token-value, semantic-bnf-token-rule-matchlist):
New API macros/functions.

(semantic-bnf-matching-to-lalr, semantic-bnf-terminal-token-p,
semantic-bnf-find-terminals, semantic-bnf-find-terminal-symbols,
semantic-bnf-to-lalr, semantic-bnf-find-parser-mode): New functions to
generate tables for the LALR parser.

(semantic-bnf-token-table, semantic-bnf-find-token-destination): New
functions to generate a table of %token elements.

(semantic-bnf-generate-and-load): Take into account new %tokentable
and %parsermode declarations.  Added generation of tables for the LALR
parser.

(semantic-bnf-mode-keywords): Correctly highlight rule name when the
colon is on a new line.

(semantic-bnf-mode): Added specification of the
`fill-paragraph-function'.

(semantic-bnf-previous-colon-indentation): Improved indentation of
rules when the colon and the matching list are on new lines.

(semantic-bnf-fill-paragraph): New function to fill BNF comments.