(semantic-bnf-token-table): Added declaration of local variable tsymb.
(semantic-beginning-of-body): New function to move the point to the
beginning of a function body.

(semantic-bnf-find-setup-code): Use `semantic-beginning-of-body' to
insert new setup code at the beginning of the setup function.

(semantic-bnf-find-table-destination-old,
semantic-bnf-find-table-destination,
semantic-bnf-find-keyword-destination,
semantic-bnf-find-token-destination): Fixed "def[var|const]" regular
expression to ensure that the searched variable or constant name is on
word boundaries.